requirements.txt
streamlit
pandas
mysql-connector-python
pymysql
sqlalchemy
scikit-learn
altair